+++ node/directives/index.ts
@@ -4,8 +4,10 @@ import { WithUserPermissions } from './withUserPermissions'
 import { CheckAdminAccess } from './checkAdminAccess'
 import { CheckUserAccess } from './checkUserAccess'
 import { AuditAccess } from './auditAccess'
+import { CheckAccessWithFeatureFlag } from './checkAccessWithFeatureFlag'
 
 export const schemaDirectives = {
+  checkAccessWithFeatureFlag: CheckAccessWithFeatureFlag as any,
   checkAdminAccess: CheckAdminAccess as any,
   checkUserAccess: CheckUserAccess as any,
   withSession: WithSession,